+++ src/gamspy/_algebra/operation.py
@@ -622,13 +622,13 @@ class Ord(operable.Operable):
 
     """
 
-    def __init__(self, set: Set | Alias):
-        if not isinstance(set, (syms.Set, syms.Alias)):
+    def __init__(self, symbol: Set | Alias):
+        if not isinstance(symbol, (syms.Set, syms.Alias)):
             raise ValidationError(
                 "Ord operation is only for Set and Alias objects!"
             )
 
-        self._set = set
+        self._symbol = symbol
         self.domain: list[Set | Alias] = []
         self.where = condition.Condition(self)
 
@@ -661,7 +661,7 @@ class Ord(operable.Operable):
         'ord(i)'
 
         """
-        return f"ord({self._set.name})"
+        return f"ord({self._symbol.name})"
 
     def latexRepr(self) -> str:
         """
@@ -671,7 +671,7 @@ class Ord(operable.Operable):
         -------
         str
         """
-        return f"ord({self._set.name})"
+        return f"ord({self._symbol.name})"
 
 
 class Card(operable.Operable):