+++ src/gamspy/_symbols/symbol.py
@@ -27,7 +27,7 @@ class Symbol:
         -------
         str
         """
-        name = self.name.replace("_", "\_")
+        name = self.name.replace("_", "\\_")
         representation = f"{name}"
         domain = list(self.domain)