+++ src/gamspy/_validation.py
@@ -404,35 +404,12 @@ def validate_model_name(name: str) -> str:
 def validate_solver_args(
     system_directory: str,
     backend: Literal["local", "engine", "neos"],
-    solver: str | None,
+    solver: str,
     problem: Problem | str,
     options: Options | None,
     output: io.TextIOWrapper | None,
     load_symbols: list[str] | None,
 ) -> None:
-    # Check validity of solver
-    if solver is not None:
-        if not isinstance(solver, str):
-            raise TypeError("`solver` argument must be a string.")
-
-        solver = solver.upper()
-        installed_solvers = utils.getInstalledSolvers(system_directory)
-        if backend == "local" and solver not in installed_solvers:
-            raise ValidationError(
-                f"Provided solver name `{solver}` is not installed on your"
-                f" machine. Install `{solver}` with `gamspy install solver"
-                f" {solver.lower()}`"
-            )
-
-        capabilities = utils.getSolverCapabilities(system_directory)
-        if str(problem) not in capabilities[solver]:
-            raise ValidationError(
-                f"Given solver `{solver}` is not capable of solving given"
-                f" problem type `{problem}`. See capability matrix "
-                "(https://www.gams.com/latest/docs/S_MAIN.html#SOLVERS_MODEL_TYPES)"
-                " to choose a suitable solver"
-            )
-
     # Check validity of options
     if options is not None and not isinstance(options, Options):
         raise TypeError(
@@ -460,6 +437,39 @@ def validate_solver_args(
                 raise ValidationError(
                     f"Elements of `load_symbols` must be of type Symbol but found {elem}"
                 )
+
+    # Check validity of solver
+    if not isinstance(solver, str):
+        raise TypeError("`solver` argument must be a string.")
+
+    if backend == "neos" and solver.lower() in ["mpsge", "kestrel"]:
+        raise ValidationError(
+            f"`{solver}` is not a valid solver for NEOS Server."
+        )
+
+    if backend == "engine" and solver.lower() in ["mpsge"]:
+        raise ValidationError(
+            f"`{solver}` is not a valid solver for GAMS Engine."
+        )
+
+    if backend == "local":
+        # No need to check whether the solver is installed on client's machine for NEOS or ENGINE.
+        installed_solvers = utils.getInstalledSolvers(system_directory)
+        if solver.upper() not in installed_solvers:
+            raise ValidationError(
+                f"Provided solver name `{solver}` is not installed on your"
+                f" machine. Install `{solver}` with `gamspy install solver"
+                f" {solver.lower()}`"
+            )
+
+        capabilities = utils.getSolverCapabilities(system_directory)
+        if str(problem).upper() not in capabilities[solver.upper()]:
+            raise ValidationError(
+                f"Given solver `{solver}` is not capable of solving given"
+                f" problem type `{problem}`. See capability matrix "
+                "(https://www.gams.com/latest/docs/S_MAIN.html#SOLVERS_MODEL_TYPES)"
+                " to choose a suitable solver"
+            )
 
 
 def validate_equations(model: Model):